Moved SweeperService to atlasdb-impl-shared, added JavaDoc
