+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/sweep/SweeperServiceImplIntegrationTest.java
@@ -22,7 +22,6 @@ import org.junit.Test;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.impl.InMemoryKeyValueService;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
-import com.palantir.atlasdb.sweeperservice.SweeperService;
 
 public class SweeperServiceImplIntegrationTest extends AbstractBackgroundSweeperIntegrationTest {
     private SweeperService sweeperService;
